# examples/api/rest/openapi.yaml
openapi: 3.0.3
info:
  title: TLE Parser API
  description: |
    RESTful API for parsing and working with Two-Line Element (TLE) satellite data.

    This API provides endpoints for:
    - Parsing TLE data in various formats
    - Validating TLE data
    - Fetching TLE data from external sources
    - Calculating orbital parameters and predictions
    - Managing TLE data in databases

    ## Features
    - Comprehensive TLE parsing and validation
    - Support for batch processing
    - Real-time satellite tracking
    - Orbital calculations using SGP4/SDP4
    - Multiple output formats (JSON, XML, CSV, etc.)
    - Rate limiting and authentication
    - API versioning

  version: 1.0.0
  contact:
    name: TLE Parser Support
    url: https://github.com/gv-sh/tle-parser
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.tleparser.com/v1
    description: Production server
  - url: https://staging-api.tleparser.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
paths:
...
  /calculate/position:
    post:
      tags:
        - Calculate
      summary: Calculate satellite position
      description: Calculate satellite position at a given time using SGP4
      operationId: calculatePosition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionRequest'
      responses:
        '200':
          description: Successfully calculated position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    PositionRequest:
      type: object
      required:
        - tle
        - timestamp
      properties:
        tle:
          type: string
        timestamp:
          type: string
          format: date-time
          example: '2024-11-21T10:30:00Z'
        coordinateSystem:
          type: string
          enum: [TEME, J2000, ECEF]
          default: TEME
...
    PositionResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        position:
          type: object
          properties:
            x:
              type: number
              description: X coordinate in km
            y:
              type: number
              description: Y coordinate in km
            z:
              type: number
              description: Z coordinate in km
        velocity:
          type: object
          properties:
            x:
              type: number
              description: X velocity in km/s
            y:
              type: number
              description: Y velocity in km/s
            z:
              type: number
              description: Z velocity in km/s
        altitude:
          type: number
          description: Altitude in km
        latitude:
          type: number
        longitude:
          type: number
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: INVALID_TLE_FORMAT
        message:
          type: string
          example: Invalid TLE format
        details:
          type: object
        field:
          type: string
        timestamp:
          type: string
          format: date-time